exporting to .env

+++ helper-functions.sh
@@ -76,30 +76,45 @@ function _waitForever {
 }
 
 function _getOverriteConfigFileContent {
+    local FORMAT=php
+    if [ ! -z "$1" ]; then
+        FORMAT=$1
+    fi
+
     local PHP_ARRAY_VAR_NAME=\$_DOCKER_CONTAINER_OVERRIDE_CONFIG
-    local ENV_PREFIX=CELY-OVERRIDE-
+    local ENV_PREFIX=CELY_OVERRIDE_
     local OVERRIDES=$(env | grep -E "$ENV_PREFIX.+")
 
     local outputContent=""
     for OVERRIDE in $OVERRIDES; do
         local SHELL_ENV_NAME=$(echo $OVERRIDE | sed -r "s/=.+//")
         local SHELL_ENV_VALUE=$(echo $OVERRIDE | sed -r "s/^.+?=//")
-        local asArray=$(echo $SHELL_ENV_NAME | sed -r "s/^$ENV_PREFIX//" | sed -r "s/,/']['/g")
-        asArray="$PHP_ARRAY_VAR_NAME['$asArray']=$SHELL_ENV_VALUE;"
 
-        outputContent="$outputContent\n $asArray"
+        if [ "$FORMAT" == "php" ]; then
+            local line=$(echo $SHELL_ENV_NAME | sed -r "s/^$ENV_PREFIX//" | sed -r "s/,/']['/g")
+            line="$PHP_ARRAY_VAR_NAME['$line']=$SHELL_ENV_VALUE;"
+        elif [ "$FORMAT" == ".env" ]; then
+            local line=$(echo $SHELL_ENV_NAME | sed -r "s/^$ENV_PREFIX//")
+            line="$line=$SHELL_ENV_VALUE"
+        fi
+
+        outputContent="$outputContent\n$line"
     done
 
-    local outputTempFile=$(tempfile)
-    printf "$outputContent" > "$outputTempFile"
-    outputContent=$(sort "$outputTempFile")
-    echo "$outputContent" > "$outputTempFile"
+    if [ "$FORMAT" == "php" ]; then
+        local outputTempFile=$(tempfile)
+        printf "$outputContent" > "$outputTempFile"
+        outputContent=$(sort "$outputTempFile")
+        echo "$outputContent" > "$outputTempFile"
 
-    printf "<?php \n /* FILE AUTO GENERATED */ \n $PHP_ARRAY_VAR_NAME = [];\n" > "$outputTempFile"
-    echo "$outputContent" >> "$outputTempFile"
-    printf "\n return $PHP_ARRAY_VAR_NAME;\n" >> "$outputTempFile"
+        printf "<?php \n /* FILE AUTO GENERATED */ \n $PHP_ARRAY_VAR_NAME = [];\n" > "$outputTempFile"
+        echo "$outputContent" >> "$outputTempFile"
+        printf "\n return $PHP_ARRAY_VAR_NAME;\n" >> "$outputTempFile"
 
-    cat "$outputTempFile"
+        cat "$outputTempFile"
 
-    rm -f "$outputTempFile"
+        rm -f "$outputTempFile"
+    elif [ "$FORMAT" == ".env" ]; then
+        printf "$outputContent\n"
+    fi
 }
